:art: Initial commit to remove GcOrganisations
