+++ tests/factories/designation.js
@@ -1,7 +1,7 @@
-import FactoryGuy from 'ember-data-factory-guy';
-import './item';
+import FactoryGuy from "ember-data-factory-guy";
+import "./item";
 
-FactoryGuy.define('designation', {
+FactoryGuy.define("designation", {
   sequences: {
     id: function(num) {
       return num;
@@ -12,23 +12,23 @@ FactoryGuy.define('designation', {
     }
   },
   default: {
-    id:                 FactoryGuy.generate('id'),
-    code:               FactoryGuy.generate('code'),
-    detailType:         'StockitLocalOrder',
-    status:             'Active',
-    state:              'submitted',
-    createdAt:          '12/07/2016',
-    updatedAt:          '12/07/2016',
-    purposeDescription: 'Test',
-    gcOrganisationId:    '',
-    gc_organisation:     FactoryGuy.belongsTo('gc_organisation'),
-    submittedBy:         FactoryGuy.belongsTo('user'),
-    beneficiary:         FactoryGuy.belongsTo('beneficiary'),
-    orders_packages:     FactoryGuy.hasMany('orders_package'),
-    orders_purposes:     FactoryGuy.hasMany('orders_purpose'),
-    goodcity_requests:   FactoryGuy.hasMany('goodcity_request')
+    id: FactoryGuy.generate("id"),
+    code: FactoryGuy.generate("code"),
+    detailType: "StockitLocalOrder",
+    status: "Active",
+    state: "submitted",
+    createdAt: "12/07/2016",
+    updatedAt: "12/07/2016",
+    purposeDescription: "Test",
+    gcOrganisationId: "",
+    organisation: FactoryGuy.belongsTo("organisation"),
+    submittedBy: FactoryGuy.belongsTo("user"),
+    beneficiary: FactoryGuy.belongsTo("beneficiary"),
+    orders_packages: FactoryGuy.hasMany("orders_package"),
+    orders_purposes: FactoryGuy.hasMany("orders_purpose"),
+    goodcity_requests: FactoryGuy.hasMany("goodcity_request")
   },
   designation_with_item: {
-    item: FactoryGuy.hasMany('item')
+    item: FactoryGuy.hasMany("item")
   }
 });
